requirements.txt
# Core Platform Dependencies
torch==2.1.0
numpy==1.24.3
scipy==1.10.1
scikit-learn==1.2.2
tensorflow==2.12.0
matplotlib==3.7.1
seaborn==0.12.2
pandas==2.0.1

# World Generation
noise==1.2.2
shapely==2.0.1
networkx==3.1

# Machine Learning
transformers==4.30.2
fastai==2.7.12
optuna==3.1.1

# Web Framework
fastapi==0.95.2
uvicorn==0.22.0
pydantic==1.10.7

# Database
sqlalchemy==2.0.12
alembic==1.10.4
psycopg2-binary==2.9.6
redis==4.5.4

# Security
cryptography==40.0.2
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1

# Performance & Monitoring
locust==2.15.1
prometheus-client==0.16.0
opentelemetry-api==1.17.0
opentelemetry-sdk==1.17.0

# Minecraft Specific
mcpi==1.1.0
pymclevel==1.1.0

# Development & Testing
pytest==7.3.1
black==23.3.0
flake8==6.0.0
mypy==1.3.0
bandit==1.7.5
safety==2.3.4

# Utilities
pyyaml==6.0
click==8.1.3
rich==13.3.5
typer==0.9.0
